Test case and robustness improvements.
This incorporates a variety of changes uncovered while updating the
test program to allow it to be generated on OSX via LLVM's
cross-compilation support.

The VCG now works directly over LLVM statements so that changes to
support additional statements due to require changes to an event type.

Error processing of SMTLIB output has been updated to avoid multiple
threads interleaving output.

We now properly update the stack pointer across calls.

+++ haskell-proto/test-programs/add_1/add.c
@@ -11,7 +11,7 @@ unsigned long fib(unsigned long x) {
     }
 }
 
-long main() {
+int main() {
   long x = add(42);
   return 0;
 }
